Refactor windows event log parsing services to use it in on-demand table
